[DEV-8722] Create suppressed Legend component
